requirements.txt
fasttext==0.9.2
fitrat==0.0.9
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-tools==1.50.0
hfstol==1.2.11
kenlm==0.1
loguru==0.6.0
protobuf==4.21.9
py-hfst-ospell @ git+https://github.com/GooDeeJAY/py-hfst-ospell.git@b7799a4b58c63d499542ef13439d28bd7b2cc3d2
pybind11==2.10.4
pydantic==1.10.2
six==1.16.0
typing_extensions==4.4.0
typesense==0.21.0
razdel~=0.5.0
ctranslate2~=4.2.1
torch~=2.0.1